cayenne: Per-file row-based deletion vectors with Vortex-native streaming scan
