+++ crates/cayenne/tests/acid_compliance_test.rs
@@ -1272,10 +1272,12 @@ async fn test_multifile_duplicate_pk_in_batch_impl(fixture: TestFixture) -> Test
 
 // =============================================================================
 // Multi-File Test 8: Position-Based Delete + Insert Across Files
-// Tests that position-based deletion also works correctly with compaction
+// Tests that position-based deletion works correctly with per-file deletion vectors
 // =============================================================================
 
-async fn test_multifile_position_based_upsert_impl(fixture: TestFixture) -> TestResult<()> {
+async fn test_multifile_position_based_delete_reinsert_impl(
+    fixture: TestFixture,
+) -> TestResult<()> {
     // Create table WITHOUT primary key (uses position-based deletion)
     let schema = Arc::new(Schema::new(vec![
         Field::new("category", DataType::Utf8, false),
@@ -1283,10 +1285,10 @@ async fn test_multifile_position_based_upsert_impl(fixture: TestFixture) -> Test
     ]));
 
     let table_options = CreateTableOptions {
-        table_name: "mf_position_upsert".to_string(),
+        table_name: "mf_position_delete_insert".to_string(),
         schema: Arc::clone(&schema),
         primary_key: vec![],
-        on_conflict: None, // No PK = position-based
+        on_conflict: None, // No PK = position-based deletion
         base_path: fixture.data_path.to_string_lossy().to_string(),
         partition_column: None,
         vortex_config: cayenne::metadata::VortexConfig::default(),
@@ -1297,11 +1299,11 @@ async fn test_multifile_position_based_upsert_impl(fixture: TestFixture) -> Test
     let table = Arc::new(CayenneTableProvider::create_table(catalog, table_options).await?);
     let ctx = SessionContext::new();
     ctx.register_table(
-        "mf_position_upsert",
+        "mf_position_delete_insert",
         Arc::clone(&table) as Arc<dyn TableProvider>,
     )?;
 
-    // File 1: Initial data
+    // File 1: Initial data (3 rows)
     let batch1 = RecordBatch::try_new(
         Arc::clone(&schema),
         vec![
@@ -1310,14 +1312,15 @@ async fn test_multifile_position_based_upsert_impl(fixture: TestFixture) -> Test
         ],
     )?;
     insert_batch(&table, batch1).await?;
-    assert_eq!(get_row_count(&ctx, "mf_position_upsert").await?, 3);
+    assert_eq!(get_row_count(&ctx, "mf_position_delete_insert").await?, 3);
 
-    // Delete category B
+    // Delete category B (removes B(200) from File 1)
     delete_records(&table, col("category").eq(lit("B"))).await?;
-    assert_eq!(get_row_count(&ctx, "mf_position_upsert").await?, 2);
+    assert_eq!(get_row_count(&ctx, "mf_position_delete_insert").await?, 2);
 
-    // File 2: Insert new data including a "B" again
-    // With position-based + compaction fix, this should work correctly
+    // File 2: Insert new data including a new "B" row
+    // With per-file deletion vectors, new files have no deletions - the B(222) is a new row,
+    // completely separate from the deleted B(200). This is NOT upsert behavior.
     let batch2 = RecordBatch::try_new(
         Arc::clone(&schema),
         vec![
@@ -1327,16 +1330,17 @@ async fn test_multifile_position_based_upsert_impl(fixture: TestFixture) -> Test
     )?;
     insert_batch(&table, batch2).await?;
 
-    // Should have 4 rows: A(100), C(300), B(222), D(400)
+    // Should have 4 rows: A(100), C(300) from File 1 + B(222), D(400) from File 2
+    // Note: B(222) is a NEW row, not a replacement - there's no PK to deduplicate
     assert_eq!(
-        get_row_count(&ctx, "mf_position_upsert").await?,
+        get_row_count(&ctx, "mf_position_delete_insert").await?,
         4,
         "Should have 4 rows after position-based delete and re-insert"
     );
 
-    // Verify B has the new value
+    // Verify only one B exists (the new one with value 222)
     let df = ctx
-        .sql("SELECT value FROM mf_position_upsert WHERE category = 'B'")
+        .sql("SELECT value FROM mf_position_delete_insert WHERE category = 'B'")
         .await?;
     let results = df.collect().await?;
     let value = results
@@ -1344,11 +1348,7 @@ async fn test_multifile_position_based_upsert_impl(fixture: TestFixture) -> Test
         .and_then(|b| b.column(0).as_any().downcast_ref::<Int64Array>())
         .and_then(|a| a.values().first())
         .copied();
-    assert_eq!(
-        value,
-        Some(222),
-        "Re-inserted category B should have new value"
-    );
+    assert_eq!(value, Some(222), "New B row should have value 222");
 
     Ok(())
 }
@@ -1467,5 +1467,5 @@ test_with_backends!(test_multifile_upsert_composite_pk_impl);
 test_with_backends!(test_multifile_delete_all_readd_all_impl);
 test_with_backends!(test_multifile_interleaved_many_files_impl);
 test_with_backends!(test_multifile_duplicate_pk_in_batch_impl);
-test_with_backends!(test_multifile_position_based_upsert_impl);
+test_with_backends!(test_multifile_position_based_delete_reinsert_impl);
 test_with_backends!(test_multifile_bulk_delete_bulk_reinsert_impl);
