+++ crates/runtime/Cargo.toml
@@ -242,6 +242,7 @@ default = [
     "keyring-secret-store",
     "aws-secrets-manager",
     "sharepoint",
+    "postgres",
     "bedrock",
 ]
 delta_lake = ["data_components/delta_lake"]
